Code cleanup, fileByFeature
