+++ wykop_handler.go
@@ -16,7 +16,6 @@ const (
 	userKeyPathElement = "/userkey/"
 	appKeyPathElement  = "/appkey/"
 	accountKeyHeader   = "accountkey"
-	privateMessagePositiveAnswer = "[true]"
 )
 
 type WykopHandler struct {
@@ -84,89 +83,6 @@ func (wh *WykopHandler) GetProfile(username string) (profile Profile, wypokError
 		End()
 
 	wypokError = wh.getObjectFromJson(responseBody, &profile)
-
-	return
-}
-
-func (wh *WykopHandler) GetConversationsList() (conversationsList []Conversation, wykopError *WykopError) {
-	urlAddress := getConversationsListUrl(wh)
-
-	_, responseBody, _ := gorequest.New().Get(urlAddress).
-									Set(contentType, mediaTypeFormType).
-									Set(apiSignHeader, wh.hashRequest(urlAddress)).
-									End()
-
-	wykopError = wh.getObjectFromJson(responseBody, &conversationsList)
-
-	return
-}
-
-func (wh *WykopHandler) SendPrivateMessageTo(to *string, message *string) (succeeded bool, wykopError *WykopError) {
-	urlAddress := getSendMessageUrl(to, wh)
-
-	body := url.Values{}
-	body.Set("body", *message)
-
-	_, responseBody, _ := gorequest.New().Post(urlAddress).
-									Set(contentType, mediaTypeFormType).
-									Set(apiSignHeader, wh.hashRequest(urlAddress+body.Get("body"))).
-									Send(body).
-									End()
-	succeeded = responseBody == privateMessagePositiveAnswer
-
-	if !succeeded {
-		wykopError = wh.getObjectFromJson(responseBody, nil)
-	}
-
-	return
-}
-
-func (wh *WykopHandler) SendPrivateMessageWithEmbeddedUrlTo(to *string, message *string, embed *string) (succeeded bool, wykopError *WykopError) {
-	urlAddress := getSendMessageUrl(to, wh)
-
-	body := url.Values{}
-	body.Set("body", *message)
-	body.Set("embed", *embed)
-
-	_, responseBody, _ := gorequest.New().Post(urlAddress).
-									Set(contentType, mediaTypeFormType).
-									Set(apiSignHeader, wh.hashRequest(urlAddress+body.Get("body")+","+body.Get("embed"))).
-									Send(body).
-									End()
-	succeeded = responseBody == privateMessagePositiveAnswer
-
-	if !succeeded {
-		wykopError = wh.getObjectFromJson(responseBody, nil)
-	}
-
-	return
-}
-
-func (wh *WykopHandler) GetConversation(conversation *string) (messages []PrivateMessage, wykopError *WykopError) {
-	urlAddress := getConversationUrl(conversation, wh)
-
-	_, responseBody, _ := gorequest.New().Get(urlAddress).
-									Set(contentType, mediaTypeFormType).
-									Set(apiSignHeader, wh.hashRequest(urlAddress)).
-									End()
-
-	wykopError = wh.getObjectFromJson(responseBody, &messages)
-
-	return
-}
-
-func (wh *WykopHandler) DeleteConversation(conversation *string) (succeeded bool, wykopError *WykopError) {
-	urlAddress := getDeleteConversationUrl(conversation, wh)
-
-	_, responseBody, _ := gorequest.New().Post(urlAddress).
-									Set(contentType, mediaTypeFormType).
-									Set(apiSignHeader, wh.hashRequest(urlAddress)).
-									End()
-	succeeded = responseBody == privateMessagePositiveAnswer
-
-	if !succeeded {
-		wykopError = wh.getObjectFromJson(responseBody, nil)
-	}
 
 	return
 }
